Retrieve resource name to UUID per tenant
No tenant context was being passed when the UUID for a resource
was being fetched (based on the resource name filter).

This resulted in tenant_id to default to admin and hence all
resources were always returned resulting in ambiguity
even when the resource had a unique name within the tenant
(but not across tenants). This was fixed by passing the
tenant_id as a context to the query.

Change-Id: I1ce8817ec58995c44133496c8529a5dc1fdd9e67