Adds missing args2body and add_known_arguments
This is the first commit of 2 towards closing GBP client bug #1403417.
It adds args2body and add_known_arguments methods to the resources
whose update operations were lacking them:
* l2policy
* policy-action
* policy-classifier
This is a necessary initial step to then show the --shared option
in all create/update operations that support that attribute.

Change-Id: I1cefcb538c16165a3cd87aa261ea215861aa3b71
Partial-Bug: #1403417