Adds --shared attribute to commands missing it
This is the final commit of 2 towards closing GBP client bug #1403417.
It shows the --shared option for commands that previously
didn't show it, but should. The following resources
have been fixed (for both create and update operations):
* l2policy
* l3policy
* network-service-policy
* policy-action
* policy-classifier
* policy-rule
* policy-rule-set
* policy-target-group
Furthermore, the CLI tests have been updated to test --shared.
As an extra, all (*_all_params) test methods updated with the
new option have also been re-checked for potentially lacking
options and, in that case, updated with the missing options.

Change-Id: Ie4d88c56a2d3e7abd8b063faeac5933a96e37dde
Closes-Bug: #1403417