Support ServiceProfile resource and new attributes
Adds support for the ServiceProfile resource, for the ServiceChainNode
resource's service_profile_id attribute, and for the shared attribute
of the ServiceChainNode and ServiceChainSpec resources.

Also updates test_requirements.txt to use the current hacking version,
and disables checks that fail. These will need to addressed in a
separate patch.

Partially implements blueprint node-centric-chain-plugin

Change-Id: I62b59b878138d3d9ad5d340a0b8f7ff0b74a39a5