Friendlier option for providing dict key arguments
Several CLI commands require providing dictionary arguments. These
arguments are currently provided as:

--argument-name "key1=<value1,key2=value2,...>"

Although the value fields are optional, it requires the user to
at least provide input in the form "key=". This is not very intuitive
and often confusing.

In this patch we preseve the current convention of providing the
dictionary arguments, but we enhance the parsing logic to also accept only
keys as arguments. The values are assumed as empty strings in such cases
and the appropriate dictonary is constructed and passed to the backend. So,
the following new syntax is also supported:

--argument-name "key1<,key2,...>"

This is a backward compatible change to the CLI.

Change-Id: I0f4c15b02c151a5f4facb2881fb0c11896a7e39c
Closes-bug: 1504912