Fix providing multiple ES for External Policy
There was a bug in processing the list of External Segments
specified by the user which was causing the name of the External
Segment to be sent to the backend. List comprehension in now being
used to build the list of UUIDs for the specified External Segments.

Change-Id: I7b79fccaffdc22e5f3a5f10b485803a0ba7634ff
Closes-bug: 1505402
(cherry picked from commit a12a46bb8e9c801969b83e303e939837bf84daf6)