[apic-mapping] Segmentation Label ext for PT CLI
The patch: https://review.openstack.org/#/c/371962/ introduced a new
extension cisco_apic_gbp_segmentation_label, for the apic policy drivers.
An extension attribute: segmentation_labels, that extends the Policy
Target definition, is being introduced in that extension. The
segementation_labels attribute is a list of strings.

This patch introduces a CLI option: --segmentation-labels for the
policy_target create and update operations. This CLI option accepts
a comma separated string as the option value.

Change-Id: I7ae44111bbe06332f2a2271094ab21c437a40658